update SwiftLint, add URL constructor
