+++ 2020-spring/Kungurov_Kirill/hw2/News/.swiftlint.yml
@@ -2,11 +2,12 @@ disabled_rules:
 - explicit_acl
 - explicit_top_level_acl
 - explicit_type_interface
-- explicit_enum_raw_value
+- explicit_enum_raw_value 
 - no_extension_access_modifier
 - no_grouping_extension
-- multiline_literal_brackets
 - vertical_whitespace_between_cases
+- unowned_variable_capture
+- file_header
 
 opt_in_rules:
 - anyobject_protocol
@@ -31,8 +32,8 @@ opt_in_rules:
 - fallthrough
 - fatal_error_message
 - file_name
+- file_types_order
 - first_where
-- last_where
 - force_unwrapping
 - function_default_parameter_at_end
 - identical_operands
@@ -40,6 +41,8 @@ opt_in_rules:
 - implicitly_unwrapped_optional
 - joined_default_parameter
 - large_tuple
+- last_where
+- legacy_multiple
 - legacy_random
 - let_var_whitespace
 - literal_expression_end_indentation
@@ -47,14 +50,14 @@ opt_in_rules:
 - mark
 - modifier_order
 - multiline_arguments
-- multiline_function_chains
-- multiline_parameters
 - multiline_arguments_brackets
-- multiline_parameters_brackets
+- multiline_function_chains
 - multiline_literal_brackets
+- multiline_parameters
+- multiline_parameters_brackets
 - nimble_operator
-- number_separator
 - nslocalizedstring_key
+- number_separator
 - object_literal
 - operator_usage_whitespace
 - overridden_super_call
@@ -67,6 +70,7 @@ opt_in_rules:
 - quick_discouraged_call
 - quick_discouraged_focused_test
 - quick_discouraged_pending_test
+- reduce_into
 - redundant_nil_coalescing
 - redundant_type_annotation
 - required_enum_case
@@ -75,22 +79,21 @@ opt_in_rules:
 - sorted_first_last
 - sorted_imports
 - static_operator
-- strong_iboutlet
 - strict_fileprivate
+- strong_iboutlet
 - switch_case_on_newline
-- toggle_bool
 - todo
+- toggle_bool
 - trailing_closure
+- type_contents_order
 - unavailable_function
 - unneeded_parentheses_in_closure_argument
 - untyped_error_in_catch
+- unused_declaration
 - unused_import
-- unused_private_declaration
 - vertical_parameter_alignment_on_call
-- vertical_whitespace_closing_braces
-- vertical_whitespace_opening_braces
-- yoda_condition
 - xct_specific_matcher
+- yoda_condition
 
 excluded:
   - Pods
@@ -98,7 +101,7 @@ excluded:
 attributes:
   always_on_same_line: ["@discardableResult", "@objc", "@IBAction", "@IBDesignable"]
 
-line_length:
+line_length: 
   warning: 160
   ignores_function_declarations: true
   ignores_comments: true
@@ -106,7 +109,7 @@ line_length:
 nesting:
   type_level: 2
 
-type_body_length: 500
+type_body_length: 500 
 
 file_length: 700
 
@@ -114,14 +117,20 @@ function_parameter_count: 6
 
 warning_threshold: 1
 
+file_name:
+  suffix_pattern: "Extensions?|\\+.*"
+
+cyclomatic_complexity: 
+  ignores_case_statements: true
+
 conditional_returns_on_newline:
-  if_only: true
+  if_only: true 
 
 identifier_name:
   excluded:
     - id
     - x
     - y
+  allowed_symbols: [_]
 
-file_name:
-  excluded: [Util.swift]
+
